macOS Apple Silicon port + Vulkan/MoltenVK backend opt-in (WIP)
Outer-repo half of the macOS arm64 + Vulkan port. The PsyCross submodule
holds the new Vulkan backend itself (separate PR on its own
`vulkan-port` branch); this commit tracks that submodule head and
provides the build-system, game-side and macOS-portability fixes the
backend needs to actually link and run.

State of the port: builds clean on macOS arm64 with both renderers; on
Apple Silicon the OpenGL path also works but exhibits geometry
flicker / dropped primitives because of the GL-on-Metal compatibility
layer's relaxed sync, which motivated the Vulkan port. Driver 2 menu /
HUD / 3D world geometry render under Vulkan; remaining gaps are blend
modes, stencil, scissor, and a few PSX scratch-render paths
(GR_StoreFrameBuffer).

Build system:
- src_rebuild/premake5.lua: --renderer={opengl|vulkan} option, macOS
  Homebrew prefix detection, Vulkan headers + loader + MoltenVK +
  shaderc (statically linked for GLSL→SPIR-V at runtime), framework
  links (Metal / QuartzCore / IOSurface), and exclusion of OpenGL
  sources when Vulkan is selected.
- src_rebuild/premake5_psycross.lua: same wiring on the static-lib
  side; `removefiles` switches PsyX_render.cpp ↔ PsyX_render_vk.cpp.
- src_rebuild/premake_modules/usage/usage.lua: tolerate the modern
  premake5 `uses` field already existing as a core API (older custom
  registration would error out on premake5 >= beta8).

Game-side fixes (LP64 / arm64):
- src_rebuild/Game/driver2.h: `trap()` macro now uses
  `__builtin_trap()` instead of the x86-only `__asm__("int3")`.
- src_rebuild/Game/psyx_compat.h: OTTYPE = OT_TAG also on
  __aarch64__/__arm64__/_M_ARM64. Without this the OT array allocates
  8-byte slots while ClearOTagR/addPrim/DrawOTag treat the array as
  12-byte OT_TAG — silent corruption of every OT linked-list pointer.
- src_rebuild/Game/dr2locale.{c,h}: malloc.h guard for macOS;
  GET_GAME_TXT/GET_MISSION_TXT cast pointers via u_intptr first to
  avoid clang's "cast from pointer to smaller type loses information"
  error.
- src_rebuild/Game/C/civ_ai.c, dr2roads.c, draw.c, event.c, glaunch.c,
  mission.c, targets.c: same intptr_t-routed casts where the engine
  packed small integers into pointer-sized fields.
- src_rebuild/utils/{fs.{cpp,h}, targa.cpp, audio_source/snd_wav_cache.cpp,
  video_source/VideoPlayer.cpp}, src_rebuild/tools/font_tool/font_tool_main.cpp:
  malloc.h guards (macOS uses stdlib.h), `__unix__` guards extended
  to include `__APPLE__`, GL-only paths in VideoPlayer guarded so
  the Vulkan build doesn't drag GL symbols.

Misc:
- BUILD_MACOS.md: build instructions (deps, premake invocation, both
  renderer options).
- data/config.ini: NTSC defaults sized for the new window.
- .gitignore: keep the user's CD-ROM images (Driver2CD1/2.bin) out of
  the repo.

+++ src_rebuild/Game/C/civ_ai.c
@@ -1083,7 +1083,7 @@ int CheckChangeLanes(DRIVER2_STRAIGHT* straight, DRIVER2_CURVE* curve, int distA
 			segLen = straight->length;
 		}
 
-		newLane = currentLane + (Random2((int)straight) >> 7 & 2U) + 0xff & 0xff;
+		newLane = currentLane + (Random2((int)(intptr_t)straight) >> 7 & 2U) + 0xff & 0xff;
 		
 		if (tryToPark)
 		{
